editoast: split ModelV2 traits into separate files
